other modifications for cross platform build

+++ src/Constants.cpp
@@ -19,9 +19,9 @@ namespace GridGeom
 
     //mesh constants
     static constexpr double minimumDeltaCoordinate = 1e-14;
-    static constexpr size_t maximumNumberOfEdgesPerNode = 10;
-    static constexpr size_t maximumNumberOfNodesPerFace = 8;
-    static constexpr size_t maximumNumberOfConnectedNodes = maximumNumberOfEdgesPerNode * 4;
+    static constexpr int maximumNumberOfEdgesPerNode = 10;
+    static constexpr int maximumNumberOfNodesPerFace = 8;
+    static constexpr int maximumNumberOfConnectedNodes = maximumNumberOfEdgesPerNode * 4;
     static constexpr double minimumCellArea = 1e-12;
 
     //orthogonalization
